refactor: remove redundant deriving Typeable
GHC 9.12 auto-derives Typeable for all types, making explicit
derivations redundant and triggering -Wderiving-typeable. Remove all
occurrences and move orphan type family instances from Client.hs to
Types.hs where the type families are defined.
