+++ lib/primitive/lib/Cardano/Wallet/Primitive/NetworkId.hs
@@ -68,7 +68,6 @@ import qualified Data.Text as T
 --              addresses. Genesis file needs to be passed explicitly when
 --              starting the application.
 data NetworkDiscriminant = Mainnet | Testnet Nat
-    deriving (Typeable)
 
 class NetworkDiscriminantCheck k where
     networkDiscriminantCheck :: SNetworkId n -> Word8 -> Bool
