+++ nix/haskell.nix
@@ -192,9 +192,7 @@ CHaP: haskell-nix: nixpkgs-recent: nodePkgs: mithrilPkgs: set-git-rev: rewrite-l
               # and test suites.
               doCoverage = coverage;
 
-              # GHC 9.12: suppress 'deriving Typeable' warning (all types
-              # auto-derive Typeable, making explicit deriving redundant).
-              ghcOptions = [ "-Wno-deriving-typeable" ];
+              ghcOptions = [];
             });
           }
